# HTTP2: Write data and trailers under one lock