--- a/src/Servers/Kestrel/Core/src/Internal/Http2/Http2FrameWriter.cs
+++ b/src/Servers/Kestrel/Core/src/Internal/Http2/Http2FrameWriter.cs
@@ -198,7 +198,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
             }
         }
 
-        public ValueTask<FlushResult> WriteResponseTrailers(int streamId, HttpResponseTrailers headers)
+        public ValueTask<FlushResult> WriteResponseTrailersAsync(int streamId, HttpResponseTrailers headers)
         {
             lock (_writeLock)
             {
@@ -256,6 +256,9 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
 
         public ValueTask<FlushResult> WriteDataAsync(int streamId, StreamOutputFlowControl flowControl, in ReadOnlySequence<byte> data, bool endStream, bool firstWrite, bool forceFlush)
         {
+            // Logic in this method is replicated in WriteDataAndTrailersAsync.
+            // Changes here may need to be mirrored in WriteDataAndTrailersAsync.
+
             // The Length property of a ReadOnlySequence can be expensive, so we cache the value.
             var dataLength = data.Length;
 
@@ -286,8 +289,11 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
             }
         }
 
-        public ValueTask<FlushResult> WriteDataAndTrailersAsync(Http2Stream stream, StreamOutputFlowControl flowControl, in ReadOnlySequence<byte> data, bool firstWrite)
+        public ValueTask<FlushResult> WriteDataAndTrailersAsync(int streamId, StreamOutputFlowControl flowControl, in ReadOnlySequence<byte> data, bool firstWrite, HttpResponseTrailers headers)
         {
+            // This method combines WriteDataAsync and WriteResponseTrailers.
+            // Changes here may need to be mirrored in WriteDataAsync.
+
             // The Length property of a ReadOnlySequence can be expensive, so we cache the value.
             var dataLength = data.Length;
 
@@ -302,25 +308,21 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2
                 // https://httpwg.org/specs/rfc7540.html#rfc.section.6.9.1
                 if (dataLength != 0 && dataLength > flowControl.Available)
                 {
-                    return WriteDataAndTrailersAsyncCore(this, stream, flowControl, data, dataLength, firstWrite);
+                    return WriteDataAndTrailersAsyncCore(this, streamId, flowControl, data, dataLength, firstWrite, headers);
                 }
 
                 // This cast is safe since if dataLength would overflow an int, it's guaranteed to be greater than the available flow control window.
                 flowControl.Advance((int)dataLength);
-                WriteDataUnsynchronized(stream.StreamId, data, dataLength, endStream: false);
+                WriteDataUnsynchronized(streamId, data, dataLength, endStream: false);
 
-                stream.ResponseTrailers.SetReadOnly();
-                stream.DecrementActiveClientStreamCount();
-                return WriteResponseTrailers(stream.StreamId, stream.ResponseTrailers);
+                return WriteResponseTrailersAsync(streamId, headers);
             }
 
-            static async ValueTask<FlushResult> WriteDataAndTrailersAsyncCore(Http2FrameWriter writer, Http2Stream stream, StreamOutputFlowControl flowControl, ReadOnlySequence<byte> data, long dataLength, bool firstWrite)
+            static async ValueTask<FlushResult> WriteDataAndTrailersAsyncCore(Http2FrameWriter writer, int streamId, StreamOutputFlowControl flowControl, ReadOnlySequence<byte> data, long dataLength, bool firstWrite, HttpResponseTrailers headers)
             {
-                await writer.WriteDataAsync(stream.StreamId, flowControl, data, dataLength, endStream: false, firstWrite);
+                await writer.WriteDataAsync(streamId, flowControl, data, dataLength, endStream: false, firstWrite);
 
-                stream.ResponseTrailers.SetReadOnly();
-                stream.DecrementActiveClientStreamCount();
-                return await writer.WriteResponseTrailers(stream.StreamId, stream.ResponseTrailers);
+                return await writer.WriteResponseTrailersAsync(streamId, headers);
             }
         }