Before:

```
[01:53:40.283] Least Requests per Connection: 442625
[01:53:40.283] Most Requests per Connection: 446844
[01:53:40.283] RPS: 118,367
[01:53:40.283] Total errors: 0
[01:53:40.287] Mean latency: 4.726ms
[01:53:40.287] Max latency: 61.612ms
```

After:

```
[02:06:08.090] Least Requests per Connection: 473015
[02:06:08.090] Most Requests per Connection: 477056
[02:06:08.090] RPS: 126,438
[02:06:08.090] Total errors: 0
[02:06:08.094] Mean latency: 4.424ms
[02:06:08.095] Max latency: 64.268ms
```

It was a pretty simple change. I only messy part is passing the Http2Stream into the frame writer.

@Tratcher @halter73 @davidfowl @brianrob