@JamesNK 2020-08-14
PR feedback applied.